// src/core/types.ts
export type QueryKey = string | unknown[]

export type QueryFunction<T = unknown> = (...args: any[]) => T | Promise<T>

export type InitialDataFunction<T> = () => T | undefined

export type InitialStaleFunction = () => boolean

export type QueryKeyHashFunction = (queryKey: QueryKey) => string

export type GetPreviousPageParamFunction<TQueryFnData = unknown> = (
  firstPage: TQueryFnData,
  allPages: TQueryFnData[]
) => unknown | undefined

export type GetNextPageParamFunction<TQueryFnData = unknown> = (
  lastPage: TQueryFnData,
  allPages: TQueryFnData[]
) => unknown | undefined

export interface QueryOptions<
  TData = unknown,
  TError = unknown,
  TQueryFnData = TData
> {
  /**
   * If `false`, failed queries will not retry by default.
   * If `true`, failed queries will retry infinitely., failureCount: num
   * If set to an integer number, e.g. 3, failed queries will retry until the failed query count meets that number.
   * If set to a function `(failureCount, error) => boolean` failed queries will retry until the function returns false.
   */
  retry?: RetryValue<TError>
  retryDelay?: RetryDelayValue
  cacheTime?: number
  isDataEqual?: (oldData: unknown, newData: unknown) => boolean
  queryFn?: QueryFunction<TQueryFnData>
  queryHash?: string
  queryKey?: QueryKey
  queryKeyHashFn?: QueryKeyHashFunction
  queryFnParamsFilter?: (args: unknown[]) => unknown[]
  initialData?: TData | InitialDataFunction<TData>
  infinite?: true
  /**
   * Set this to `false` to disable structural sharing between query results.
   * Defaults to `true`.
   */
  structuralSharing?: boolean
  /**
   * This function can be set to automatically get the previous cursor for infinite queries.
   * The result will also be used to determine the value of `hasPreviousPage`.
   */
  getPreviousPageParam?: GetPreviousPageParamFunction<TQueryFnData>
  /**
   * This function can be set to automatically get the next cursor for infinite queries.
   * The result will also be used to determine the value of `hasNextPage`.
   */
  getNextPageParam?: GetNextPageParamFunction<TQueryFnData>
}